保持同步

+++ trunk/onlineTest/work/views.py
@@ -17,6 +17,7 @@ from judge.views import get_testCases
 from work.models import HomeWork, HomeworkAnswer, BanJi, MyHomework, TempHomeworkAnswer
 from django.contrib.auth.decorators import permission_required, login_required
 from django.conf import settings
+from django.db.models import Q
 from process.views import get_similarity, update_ansdb,get_similarity_v2
 import logging
 
@@ -860,6 +861,37 @@ def show_banji(request, pk):
                'title': '班级"' + banji.name + '"的信息', 'scores': students_scores}
     return render(request, 'banji_detail.html', context=context)
 
+@login_required()
+def get_students(request,banji=None):
+    json_data = {}
+    if banji!=None and banji!='':
+        banji = BanJi.objects.get(pk=banji)
+    else:
+        return HttpResponse(json.dumps(json_data))
+    users = banji.students
+    recodes = []
+    offset = int(request.GET['offset'])
+    limit = int(request.GET['limit'])
+    try:
+        qset = ( Q(username__icontains=request.GET['search']) |
+                 Q(email__icontains=request.GET['search']) |
+                 Q(id_num__icontains=request.GET['search']) )
+        users = users.filter(qset)
+    except:
+        pass
+    try:
+        sort = request.GET['sort']
+    except MultiValueDictKeyError:
+        sort = 'id_num'
+    if request.GET['order'] == 'desc':
+        sort = '-' + sort
+    json_data['total'] = users.count()
+    for user in users.all().order_by(sort)[offset:offset + limit]:
+        recode = {'username': user.username, 'pk': user.pk,
+                  'email': user.email, 'id_num': user.id_num, 'group': user.groups.all()[0].name, 'id': user.pk}
+        recodes.append(recode)
+    json_data['rows'] = recodes
+    return HttpResponse(json.dumps(json_data))
 
 @permission_required('work.change_banji')
 def add_students(request, pk):
@@ -888,6 +920,39 @@ def ajax_add_students(request):
     banji.students.add(student)
     return HttpResponse(json.dumps({'result': 0, 'count': 1}))
 
+def reset_stupassword(request):
+    if request.method == 'POST':
+        stu_id = request.POST.get('id')
+        banji_id = request.POST.get('banji_id')
+        try:
+            student = MyUser.objects.get(pk=stu_id)
+            banji = BanJi.objects.get(pk=banji_id)
+            if banji.teacher==student:
+                return HttpResponse(2)
+            stu_nu = student.id_num
+            student.set_password(stu_nu)
+            student.save()
+        except:
+            return HttpResponse(0)
+        return HttpResponse(1)
+    else:
+        return HttpResponse(0)
+
+def del_students(request):
+    if request.method == 'POST':
+        stu_id = request.POST.get('id')
+        banji_id = request.POST.get('banji_id')
+        try:
+            student = MyUser.objects.get(pk=stu_id)
+            banji = BanJi.objects.get(pk=banji_id)
+            if banji.teacher==student:
+                return HttpResponse(2)
+            banji.students.remove(student)
+        except:
+            return HttpResponse(0)
+        return HttpResponse(1)
+    else:
+        return HttpResponse(0)
 
 @permission_required('work.add_homework')
 def assign_homework(request):
@@ -1087,6 +1152,8 @@ def get_gaicuo_score(homework_answer, judged_score=0):
 
 @login_required()
 def list_finished_homework(request):
+    if not request.user.isTeacher() and not request.user.is_admin:
+        raise PermissionDenied
     banjis = BanJi.objects.filter(students=request.user).all()
     return render(request, 'finidshed_homework_list.html',
                   context={'classnames': banjis, 'position': 'finished', 'title': '查看作业结果'})
@@ -1099,6 +1166,8 @@ def get_finished_homework(request):
     :param request: 请求
     :return: 含有用户已完成作业的json
     """
+    if not request.user.isTeacher() and not request.user.is_admin:
+        raise PermissionDenied
     json_data = {}
     records = []
     user = request.user
@@ -1257,6 +1326,8 @@ def list_coursers(request):
     """
     列出课程
     """
+    if not request.user.is_admin:
+        raise PermissionDenied 
     coursers = ClassName.objects.all()
     return render(request, 'courser_list.html', {'coursers': coursers, 'title': '课程列表', 'position': 'courser_manage'})
 
@@ -1277,6 +1348,8 @@ def list_kp2s(request, id):
 
 @permission_required('judge.delete_classname')
 def delete_courser(request):
+    if not request.user.is_admin:
+        raise PermissionDenied
     try:
         ClassName.objects.get(id=request.POST['id']).delete()
         return HttpResponse(1)
@@ -1286,6 +1359,8 @@ def delete_courser(request):
 
 @permission_required('judge.delete_knowledgepoint1')
 def delete_kp1(request):
+    if not request.user.is_admin:
+        raise PermissionDenied
     try:
         KnowledgePoint1.objects.get(id=request.POST['id']).delete()
         return HttpResponse(1)
@@ -1294,6 +1369,8 @@ def delete_kp1(request):
 
 @permission_required('judge.delete_knowledgepoint2')
 def delete_kp2(request):
+    if not request.user.is_admin:
+        raise PermissionDenied
     try:
         KnowledgePoint2.objects.get(id=request.POST['id']).delete()
         return HttpResponse(1)
@@ -1303,6 +1380,8 @@ def delete_kp2(request):
 
 @permission_required('judge.add_knowledgepoint1')
 def add_kp1(request):
+    if not request.user.is_admin:
+        raise PermissionDenied
     kp1 = KnowledgePoint1(name=request.POST['name'], classname_id=request.POST['id'])
     kp1.save()
     return HttpResponse(1)
@@ -1310,6 +1389,8 @@ def add_kp1(request):
 
 @permission_required('judge.add_knowledgepoint2')
 def add_kp2(request):
+    if not request.user.is_admin:
+        raise PermissionDenied
     kp2 = KnowledgePoint2(name=request.POST['name'], upperPoint_id=request.POST['id'])
     kp2.save()
     return HttpResponse(1)
